REF: de-duplicate check_reduce_frame
